refactor: 테스트 공용으로 사용되는 이벤트값들 파일분리
